LayoutTests/fast/dom/mutation-events-disablement-expected.txt

This tests disabling mutation events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mutationEvents is []
PASS mutationEvents is []
PASS successfullyParsed is true

TEST COMPLETE
world
